Cleaned up Screens
- Moved contents of show() methods into constructors for those screens.

Now, we don't need to keep a 'wasHidden' attribute for each screen.

Previously, tables, box2D worlds, cameras, etc. were instantiated in the show() method of the screens.

Game.setScreen( screen ) calls screen.show(). So, without a 'wasHidden' variable, the objects would be instantiated again and again which lead to multiple copies being shown on the screen.

A 'wasHidden' variable was used to prevent this - when show() was called for the very first time, all the objects were instantiated. When Game.setScreen() was called to switch this screen with another screen, hide() was called on the current screen which set wasHidden to true. When Game.setScreen() was called to switch back to this screen, show() was called but since wasHidden was true, show() returned early and didn't instantiate the objects again.

Now, we've moved all such object instantiation into the constructor. There is no need to use wasHidden now since the constructor is called only once (screens are singleton).

- For similar reasons, we historically had the 'wasShown' attribute in LevelScreen which has now been removed.

- LevelRenderer.show() doesn't called LevelScreen.show() because the latter is empty.